requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0

# Deep Learning
torch==2.1.0
torch-geometric==2.4.0

# Recommendation Systems
recbole==1.2.0
scikit-surprise==1.1.3

# Machine Learning
xgboost==2.0.0
lightgbm==4.1.0

# Time Series
statsmodels==0.14.0
prophet==1.1.5
pmdarima==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Dashboard
streamlit==1.27.0

# Experiment Tracking
mlflow==2.7.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# Jupyter
jupyter==1.0.0
notebook==7.0.3
ipywidgets==8.1.0

# Testing
pytest==7.4.2
